+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 ## [Unreleased]
 
+## [0.4.5] - 2022-08-25
+
+### Fixes
+
+- Fixes the error messages when you finish current task.
+They are more descriptive, see [#56](https://github.com/TheSinding/taskstarter/issues/56)
+
 ## [0.4.4] - 2022-08-25
 
 ### Fixes
 
-- Fies current task not being set when a user reloads the window
+- Fixes current task not being set when a user reloads the window
 
 ## [0.4.3] - 2022-08-20